:wrench:/:electron: adjust work units, add datasets and tags
Adjust the work units cli to use unit as the entry command, and add
instead of register.

Slightly adjust the project structure and update the initialisation
process.

Add a dataset cli for registering new datasets. This will need further
fleshing out, but in the future the idea is that remote datasets can be
offloaded and pmtk will provide the ability to save just a metadata
config file to a git repository, and then sync the datasets from the
remote repositories if needed. This ensure a lightweight git repository
for moving the project around, and makes handling datasets clean.

Add project tagging to help track themes in a project.